Refactor component imports and clean up unused files
- Updated import paths in App component for better organization.
- Reformatted ESLint configuration for consistency.
- Removed unused Background, Header, Hero, Login, Logout, and Profile components to streamline the codebase.
