+++ src/components/auth/Login.tsx
@@ -1,4 +1,4 @@
-import { useAuth0 } from "@auth0/auth0-react";
+import { useAuth0 } from '@auth0/auth0-react';
 
 const LoginButton = () => {
   const { loginWithRedirect, isAuthenticated } = useAuth0();
@@ -7,8 +7,7 @@ const LoginButton = () => {
     !isAuthenticated && (
       <button
         onClick={() => loginWithRedirect()}
-        className="px-4 py-2  hover:underline text-white font-semibold rounded-lg shadow transition-colors duration-200"
-      >
+        className="px-4 py-2  hover:underline text-white font-semibold rounded-lg shadow transition-colors duration-200">
         Log In
       </button>
     )
