+++ src/components/auth/Profile.tsx
@@ -1,4 +1,4 @@
-import { useAuth0 } from "@auth0/auth0-react";
+import { useAuth0 } from '@auth0/auth0-react';
 
 const Profile = () => {
   const { user, isAuthenticated, isLoading } = useAuth0();
