+++ src/components/layout/Header.tsx
@@ -1,14 +1,12 @@
-import Profile from "./profile";
-import LoginButton from "./login";
-import LogoutButton from "./logout";
+import Profile from '../auth/Profile';
+import LoginButton from '../auth/Login';
+import LogoutButton from '../auth/Logout';
 
 export function Header() {
   return (
     <header className="top-0 z-20 w-full absolute flex items-center px-8 py-3">
-      {/* Logo y nombre */}
-      {/* Botones y perfil */}
       <div className="flex flex-1 items-center justify-end gap-4 ">
-        <LoginButton /> 
+        <LoginButton />
         <LogoutButton />
         <Profile />
       </div>
