+++ src/main.tsx
@@ -1,22 +1,21 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import "./index.css";
-import App from "./App.tsx";
-import { Auth0Provider } from "@auth0/auth0-react";
-import { AUTH0_CLIENT_ID, AUTH0_DOMAIN } from "./utils/constants.ts";
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import './index.css';
+import App from './App.tsx';
+import { Auth0Provider } from '@auth0/auth0-react';
+import { AUTH0_CLIENT_ID, AUTH0_DOMAIN } from './utils/constants.ts';
 
 const domain = AUTH0_DOMAIN;
 const clientId = AUTH0_CLIENT_ID;
 
-createRoot(document.getElementById("root")!).render(
+createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <Auth0Provider
       domain={domain}
       clientId={clientId}
       authorizationParams={{
         redirect_uri: window.location.origin,
-      }}
-    >
+      }}>
       <App />
     </Auth0Provider>
   </StrictMode>
